Commiting the motion vector kernel we worked on today
The kernel compiles but we have not yet testing setting it up through
host code to run it. So not sure if it works. Also, I realize the memory
usage will be crap and we will also have a tiny work group size the way
we set it up. I figured out a much better way which I will hopefully get
to implement before the deadline.